+++ src/bas_air_unit_network_dataset/models/waypoint.py
@@ -41,6 +41,7 @@ class Waypoint:
             "colocated_with": "str",
             "last_accessed_at": "date",
             "last_accessed_by": "str",
+            "elevation_ft": "int",
             "comment": "str",
         },
     }
@@ -51,6 +52,7 @@ class Waypoint:
         "colocated_with": "str",
         "last_accessed_at": "date",
         "last_accessed_by": "str",
+        "elevation_ft": "int",
         "comment": "str",
     }
 
@@ -63,6 +65,7 @@ class Waypoint:
         colocated_with: Optional[str] = None,
         last_accessed_at: Optional[date] = None,
         last_accessed_by: Optional[str] = None,
+        elevation_ft: Optional[int] = None,
         comment: Optional[str] = None,
     ) -> None:
         """
@@ -95,6 +98,7 @@ class Waypoint:
         self._colocated_with: Optional[str] = None
         self._last_accessed_at: Optional[date] = None
         self._last_accessed_by: Optional[str] = None
+        self._elevation_ft: Optional[int] = None
         self._comment: Optional[str] = None
 
         if identifier is not None:
@@ -117,6 +121,9 @@ class Waypoint:
 
         self._last_accessed_at = last_accessed_at
         self._last_accessed_by = last_accessed_by
+
+        if elevation_ft is not None:
+            self._elevation_ft = elevation_ft
 
         if comment is not None:
             self.comment = comment
@@ -327,6 +334,24 @@ class Waypoint:
         self._last_accessed_by = last_accessed_by
 
     @property
+    def elevation_ft(self) -> Optional[int]:
+        """
+        Waypoint elevation in feet.
+
+        Returns `None` if unknown. Values are positive integers.
+        """
+        return self._elevation_ft
+
+    @elevation_ft.setter
+    def elevation_ft(self, elevation_ft: int) -> None:
+        """Set waypoint elevation in feet."""
+        if elevation_ft < 0:
+            msg = "Elevation must be a positive integer."
+            raise ValueError(msg)
+
+        self._elevation_ft = elevation_ft
+
+    @property
     def comment(self) -> Optional[str]:
         """
         Waypoint comment.
@@ -373,6 +398,8 @@ class Waypoint:
         ):
             self.last_accessed_at = date.fromisoformat(feature["properties"]["last_accessed_at"])
             self.last_accessed_by = feature["properties"]["last_accessed_by"]
+        if feature["properties"]["elevation_ft"] is not None:
+            self.elevation_ft = feature["properties"]["elevation_ft"]
         if feature["properties"]["comment"] is not None:
             self.comment = feature["properties"]["comment"]
 
@@ -386,7 +413,7 @@ class Waypoint:
         self.identifier = gpx_waypoint.name
         self.geometry = [gpx_waypoint.longitude, gpx_waypoint.latitude]
 
-        if gpx_waypoint.description is None or gpx_waypoint.description == "N/A | N/A | N/A | N/A | N/A":
+        if gpx_waypoint.description is None or gpx_waypoint.description == "N/A | N/A | N/A | N/A | N/A | N/A | N/A":
             pass
 
         comment_elements = gpx_waypoint.description.split("|")
@@ -395,6 +422,7 @@ class Waypoint:
         last_accessed_at = comment_elements[2].strip()
         last_accessed_by = comment_elements[3].strip()
         comment = comment_elements[4].strip()
+        elevation_ft = comment_elements[5].strip()
 
         if name != "N/A":
             self.name = name
@@ -404,6 +432,8 @@ class Waypoint:
             self.last_accessed_at = date.fromisoformat(last_accessed_at)
         if last_accessed_by != "N/A":
             self.last_accessed_by = last_accessed_by
+        if elevation_ft != "N/A":
+            self.elevation_ft = int(elevation_ft)
         if comment != "N/A":
             self.comment = comment
 
@@ -480,6 +510,10 @@ class Waypoint:
         if self.last_accessed_by is not None:
             last_accessed_by = self.last_accessed_by
 
+        elevation_ft = "-"
+        if self.elevation_ft is not None:
+            elevation_ft = self.elevation_ft
+
         comment = "-"
         if self.comment is not None:
             comment = self.comment
@@ -494,6 +528,7 @@ class Waypoint:
             "longitude_ddm": geometry_ddm["lon"],
             "last_accessed_at": last_accessed_at,
             "last_accessed_by": last_accessed_by,
+            "elevation_ft": elevation_ft,
             "comment": comment,
         }