+++ src/bas_air_unit_network_dataset/models/waypoints.py
@@ -114,6 +114,7 @@ class WaypointCollection:
                 "longitude_dd",
                 "last_accessed_at",
                 "last_accessed_by",
+                "elevation_ft",
                 "comment",
             ]
         if inc_ddm_lat_lon:
@@ -125,6 +126,7 @@ class WaypointCollection:
                 "longitude_ddm",
                 "last_accessed_at",
                 "last_accessed_by",
+                "elevation_ft",
                 "comment",
             ]
         if inc_dd_lat_lon and inc_ddm_lat_lon:
@@ -138,6 +140,7 @@ class WaypointCollection:
                 "longitude_ddm",
                 "last_accessed_at",
                 "last_accessed_by",
+                "elevation_ft",
                 "comment",
             ]