+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+* Dependencies updated
+  [#220](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/220)
+
 ## [0.4.0] - 2023-12-03 [BREAKING!]
 
 ### Removed [BREAKING!]