+++ CHANGELOG.md
@@ -9,9 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* waypoint fuel attribute
+* Waypoint fuel attribute
   [#221](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/221)
-* waypoint elevation (ft) attribute
+* Waypoint elevation (ft) attribute
   [#222](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/222)
 * Waypoint category attribute
   [#225](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/225)